pyproject.toml
# package build
# https://peps.python.org/pep-0517/
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[tool.setuptools_scm]

# package basics
# https://peps.python.org/pep-0621/
[project]
name = "speakerbox-analysis"
description = "Data processing and analysis functions for the Speakerbox Demographics and Interruptions paper."
keywords = []
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT License" }
authors = [
  { email = "evamaxfieldbrown@gmail.com", name = "Eva Maxfield Brown" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Natural Language :: English",
  "Programming Language :: Python :: 3.10",
]
dynamic = ["version"]
dependencies = [
  "cdp-backend[pipeline]==3.1.3",
  "cdp-data==0.0.7",
  "dataclasses-json",
  "fire==0.4.0",
  "GitPython==3.1.27",
  "pandas==1.4.2",
  "pyarrow==8.0.0",
  "quilt3==5.0.0",
  "s3fs[boto3]==2022.7.1",
  "speakerbox==0.2.0",
  "tqdm==4.64.0",
]

[project.urls]
Homepage = "https://github.com/PugetSoundClinic-PIT/speakerbox-analysis"
"Bug Tracker" = "https://github.com/PugetSoundClinic-PIT/speakerbox-analysis/issues"
Documentation = "https://PugetSoundClinic-PIT.github.io/speakerbox-analysis"
"User Support" = "https://github.com/PugetSoundClinic-PIT/speakerbox-analysis/issues"

# extra dependencies
# https://peps.python.org/pep-0621/#dependencies-optional-dependencies
[project.optional-dependencies]
lint = [
  "black>=22.3.0",
  "check-manifest>=0.48",
  "flake8>=3.8.3",
  "flake8-debugger>=3.2.1",
  "flake8-pyprojecttoml",
  "flake8-typing-imports>=1.9.0",
  "isort>=5.7.0",
  "mypy>=0.790",
  "pre-commit>=2.20.0",
]
test = [
  "coverage>=5.1",
  "pytest>=5.4.3",
  "pytest-cov>=2.9.0",
  "pytest-raises>=0.11",
]
docs = [
  # Sphinx + Doc Gen + Styling
  "m2r2>=0.2.7",
  "Sphinx>=4.0.0",
  "furo>=2022.4.7",
  # Extensions
  "numpydoc",
  "sphinx-copybutton",
  # TODO: Pins
  "docutils>=0.18,<0.19",
]
dev = [
  "ipython>=8.4.0",
]

# entry points
# https://peps.python.org/pep-0621/#entry-points
[project.entry-points."console_scripts"]
speakerbox-analysis = "speakerbox_analysis._bin.speakerbox_analysis:main"

# build settings
# https://setuptools.pypa.io/en/latest/userguide/pyproject_config.html
[tool.setuptools]
zip-safe = false
include-package-data = true

[tool.setuptools.packages.find]
exclude = ["*docs/*", "*tests/*"]

[tool.setuptools.package-data]
"*" = ["*.yaml", "py.typed"]

# tools
[tool.black]
line-length = 88

[tool.isort]
ensure_newline_before_comments = true
force_grid_wrap = 0
include_trailing_comma = true
line_length = 88
multi_line_output = 3
profile = "black"
use_parentheses = true

# https://github.com/mgedmin/check-manifest#configuration
[tool.check-manifest]
ignore = [
  ".editorconfig",
  ".pre-commit-config.yaml",
  "CODE_OF_CONDUCT.md",
  "CONTRIBUTING.md",
  "Justfile",
  ".cookiecutter.yaml",
  "*docs/*",
  "*docker/*",
  "*terraform/*",
]

[tool.mypy]
files = "speakerbox_analysis/*.py"
ignore_missing_imports = true
disallow_untyped_defs = true
check_untyped_defs = true
show_error_codes = true

# https://flake8.pycqa.org/en/latest/user/options.html
# https://gitlab.com/durko/flake8-pyprojecttoml
[tool.flake8]
max-line-length = 88
ignore = "E203,E402,W291,W503"
min-python-version = "3.10.0"
per-file-ignores = [
  "__init__.py:F401",
]